+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 * Increased ASCII strings performance.
 * Improved property look up performance.
   Should now also save the property order, however, indexes are not handled somehow special.
+* Added `js_resolvetypename` to fetch proper instance type name from the constructor.